backend/requirements.txt
Flask
Flask-JWT-Extended
Werkzeug
mysql-connector-python
python-dotenv
Flask-Cors
